#5: PersistenceTest moved to source part
